gpt-neox : initial load and eval in ggml
Only tested with Pythia, and for now it seems only the first few
completed words are okay, and then it degrades.

Still committing this, because I want to be able to retrospectively
see the progress I made.  This is the cumulation of 3 to 4 days of work,
but I unfortunately did not commit my work before now.